requirements.txt
streamlit
pandas
pillow
pdf2image
pytesseract
openai
python-dotenv